src/PHPVersion/requirements.txt
Pillow==9.3.0
pillow_heif==0.7.0
pywin32==304
